\name{MPAT-package}
\alias{MPAT-package}
\alias{MPAT}
\docType{package}
\title{
Multiple Phenotype Genetic Association Testings by Combining Univariate Testing Summary Statistics.
}
\description{
This package is for conducting multiple phenotype genetic association testings by combining univariate summary statistics for each phentype in GWAS. 
It contains fourteen methods that can be used in a variety of situations. An overall p-value will be returned for assessing the significance of associations between 
a SNP and multiple phenotypes. It contains the following functions: PC, PCMinP, PCFisher, PCLC, PCSS, MinP, WI, VC, DSUM, SUM, Wald, mixAda, mixFisher, 
mixTippett, mixVar, mixSD.
}
\details{
\tabular{ll}{
Package: \tab MPAT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-11\cr
License: \tab License: GPL(>= 2)\cr
}

}
\author{
Zhonghua Liu <zliu@mail.harvard.edu>

Maintainer: Zhonghua Liu <zliu@mail.harvard.edu>

}
\references{
MPAT: an R package for multivariate phenotype association studies.
}

